feat(simulation): move ekubo_v3 filter to filters mod
