+++ crates/tycho-simulation/src/evm/protocol/filters.rs
@@ -1,5 +1,8 @@
+use revm::primitives::Address;
 use tracing::{debug, info};
 use tycho_client::feed::synchronizer::ComponentWithState;
+
+use crate::evm::protocol::ekubo_v3::extension_type;
 
 /// Filters out pools that DCI currently fails to find some accounts for
 pub fn balancer_v2_pool_filter(component: &ComponentWithState) -> bool {
@@ -110,4 +113,16 @@ pub fn erc4626_filter(component: &ComponentWithState) -> bool {
         return false;
     }
     true
+}
+
+/// Filters out unsupported ekubo_v3 extensions.
+pub fn ekubo_v3_extension_filter(component: &ComponentWithState) -> bool {
+    component
+        .component
+        .static_attributes
+        .get("extension")
+        .is_some_and(|extension_bytes| {
+            Address::try_from(&extension_bytes[..])
+                .is_ok_and(|extension| extension_type(extension).is_some())
+        })
 }
